refactor: co-locate context
