+++ site/src/pages/CreateWorkspacePage/CreateWorkspaceExperimentRouter.tsx
@@ -1,7 +1,7 @@
 import { templateByName } from "api/queries/templates";
 import { ErrorAlert } from "components/Alert/ErrorAlert";
 import { Loader } from "components/Loader/Loader";
-import { ExperimentalFormContext } from "contexts/ExperimentalFormContext";
+import { ExperimentalFormContext } from "./ExperimentalFormContext";
 import { useDashboard } from "modules/dashboard/useDashboard";
 import type { FC } from "react";
 import { useQuery } from "react-query";
