+++ site/src/pages/CreateWorkspacePage/CreateWorkspacePageView.tsx
@@ -26,7 +26,7 @@ import { Spinner } from "components/Spinner/Spinner";
 import { Stack } from "components/Stack/Stack";
 import { Switch } from "components/Switch/Switch";
 import { UserAutocomplete } from "components/UserAutocomplete/UserAutocomplete";
-import { ExperimentalFormContext } from "contexts/ExperimentalFormContext";
+import { ExperimentalFormContext } from "./ExperimentalFormContext";
 import { type FormikContextType, useFormik } from "formik";
 import { generateWorkspaceName } from "modules/workspaces/generateWorkspaceName";
 import {
