+++ site/src/pages/CreateWorkspacePage/CreateWorkspacePageViewExperimental.tsx
@@ -13,7 +13,7 @@ import { Spinner } from "components/Spinner/Spinner";
 import { Stack } from "components/Stack/Stack";
 import { Switch } from "components/Switch/Switch";
 import { UserAutocomplete } from "components/UserAutocomplete/UserAutocomplete";
-import { ExperimentalFormContext } from "contexts/ExperimentalFormContext";
+import { ExperimentalFormContext } from "./ExperimentalFormContext";
 import { type FormikContextType, useFormik } from "formik";
 import { useDebouncedFunction } from "hooks/debounce";
 import { ArrowLeft, CircleAlert, TriangleAlert } from "lucide-react";
